🐛 fix(법령검증·증여의제): 개정 감시가 낡은 본문을 보던 경로 수정 + §34의3·§34의5 등재 + RangeError 500 차단 (W15)
증여의제 §45의3·§45의5 코드리뷰 수정 1단계. 「인프라 — 개정 감시·런타임 예외·문서」.

## 1. `verify:legal`이 **공포본**을 보고 있었다 (신규 발견)

`fetchArticle`은 `lawSearch.do`가 준 MST로 `lawService.do?target=law&MST=`를 호출한다.
이건 **그 법령본이 공포된 시점의 본문**이라 두 방향으로 어긋난다 — 2026-09-18 실측:

| 법령 | 공포본(MST) | 현행 시행본(eflaw) |
|---|---|---|
| 상증법 §45의5 | 「지배주주**의** 특수관계인과」 (구) | 「지배주주 **및 그** 특수관계인과」 |
| 소득세법 시행령 §178의8 | 2026.2.27. 공포분 ②항 포함 (**미시행**) | 한 문장뿐 |

매니페스트 31개 법령 중 **9개**가 두 경로의 시행일자가 달랐다(소득세법 20260101→20260701,
지방세법 동일, 부가가치세법 시행령 20260227→20260401 등). eflaw가 미래(시행예정) 본문을
준 경우는 **0건**이었다.

⇒ `fetchArticle`에 `lawId`를 넘기면 `target=eflaw&ID=`(시행일자별)로 현행 시행본을 받는다.
   실패·빈 응답이면 공포본으로 되돌아간다. 캐시 키는 경로별로 가른다.
⇒ 「어느 본문을 기준으로 검증하는가」를 규칙에 명시하는 `basis` 축을 신설했다
   (`"enforced"` 기본 / 시행예정 조문을 코드가 선반영한 규칙만 `"announced"`).

### 경로를 바꾸자 드러난 것 — 종전엔 전부 초록이었다

- **증권거래세 탄력세율**: 영 제36001호(2026.1.1. 시행)로 코스피 5 / 코넥스 10 /
  코스닥·K-OTC 20이 됐는데 키워드는 「1만분의 15」였다. **엔진 매트릭스는 이미 정확**해
  세액 영향은 없다 — 감시 쪽만 낡아 있었다. 세율 3종을 전부 키워드로 건다.
- **국외전출세 §178의8·§178의9**: 키워드가 아직 **시행되지 않은** 2026.2.27. 공포분
  문언이었다. 코드가 그 조문을 선반영했으므로(`ExitTaxHoldingReportForm.tsx:186`)
  `basis: "announced"`로 의도를 명시했다. 시행되면 `"enforced"`로 되돌릴 것.

## 2. 상증령 §34의3·§34의5가 감시 밖이었다 (RC-R · SC-R)

§45의3·§45의5 엔진이 계산하는 실체는 전부 시행령인데 `legal-codes`에는 본법만 있었다.
커버리지 수집기는 **문자열 값 leaf만** 모으므로 JSDoc 주석의 시행령은 모수에 안 들어간다.
⇒ `GIFT.RELATED_CORP_DECREE`·`SPECIFIC_CORP_DECREE`를 상수 **값**으로 올리고
   `additions-inheritance-decree.ts`에 등재했다(361 → 363건).

키워드는 **엔진 상수를 담은 문언**으로 골랐다 — 조문 제목·용어만 넣으면 개정돼도 통과한다.
실제로 상증법 §45의5 규칙의 종전 키워드 3개는 2026.1.1. 거래상대방 확대 개정에 대해
**구별력이 0**이었다(개정 전후 본문에 셋 다 그대로 있었다).

mutation probe 11건 전건 기대대로 — 정상거래비율 40→45, 한계보유비율 3→4,
1억원→2억원, 현저대가 30%→20%, 탄력세율 5→3, 개정 전 문언 복원, basis 뒤집기 모두 FAIL.

## 3. 소수 분모가 API를 500으로 죽였다 (SC-P)

`safeMultiplyThenDivide`의 BigInt 분기는 `Math.floor`로 피연산자를 절사한다.

| 입력 | 종전 | 올바른 값 |
|---|---|---|
| `(1e10, 1e6, 0.5)` | **RangeError: Division by zero** | 20000000000000000 |
| `(1e10, 1e6, 100.25)` | 100000000000000 (0.25% 과대) | 99750623441396 |

`c === 0` 가드는 엄격 동등 비교라 0<|c|<1을 못 막았고, ⑫ `ratioSchema`가 소수 분모를
허용해 `/api/calc/gift-deemed`가 HTTP 500으로 떨어졌다(재현 확인).

⇒ BigInt 분기는 **세 피연산자가 모두 정수일 때만** 타고, 아니면 부동소수 경로로 되돌린다.
⇒ ⑫에서 `ratioSchema`(6개 유형 공유)와 roster `shares`/`totalShares`를 정수로 강제했다.
   생산 측 `parseRatio`는 언제나 `{Math.round(pct×100), 10_000}`이라 정상 입력은 안 막힌다.

회귀 테스트 4건 추가, 가드 제거 뮤테이션에서 4건 전부 실패함을 확인했다.
※ 같은 절사가 `safeMultiply`에도 있으나 나눗셈이 없어 예외로 죽지 않아 범위 밖으로 뒀다.

## 4. 인용 드리프트 (RC-S 일부 · XX-B)

- `related-corp.ts` 결과 note의 「§45의3②⑲」 — 법 §45의3은 ①~⑤가 전부다. ⑲는 **시행령**
  §34의3⑲이고 내용도 «증여자 특정»이지 출자관계 합산 근거가 아니다. 이 문자열은
  결과 화면에 그대로 렌더된다. → 「상증법 §45의3② · 상증령 §34의3⑬」.
- `router.ts`의 「Phase 3 router 후처리」 — 그런 후처리는 없다(switch 하나가 전부).
- `dup-exclusion.ts`의 §43① 열거 「§33~§45의5」 — 법문은 §45의2를 **제외**하고 §44는 포함한다.
  verbatim으로 교체하고 프로덕션 호출처 0건임을 명시했다.

## 검증

- `npm run verify:legal` 363건 / 실패 0 / 오류 0
- `npx tsc --noEmit` 0건 · `npm run lint` 경고 343건(master와 동일, 신규 0)
- `npm test` 2079파일 21749테스트 통과 (leaf는 321곳에서 쓰인다)

## 이 커밋에서 뺀 것

같은 클러스터(RC-S)의 두 건은 **같은 파일을 다시 여는 뒤 작업으로 넘긴다** —
- RC-2-f(간접출자 경유 3개 이상 1원 오차) → W10이 `computeIndirectRatio`를 구조 변경한다.
- RC-3-h(⑩1호를 중견·일반이 선택 가능) → W9이 매출처 입력 구조를 새로 만든다.

+++ lib/validators/gift-deemed-input.ts
@@ -113,7 +113,15 @@ const freeLoanAggregatedSchema = z.object({
 });
 
 // ── Phase 2: 자본거래 (평가가액·주식수 직접 입력) — sub-case 필드는 caseType별 optional ──
-const ratioSchema = z.object({ numer: z.number().nonnegative(), denom: z.number().positive() });
+// 🔴 `.int()`는 장식이 아니다 — 소수 분모가 통과하면 엔진의 `safeMultiplyThenDivide`가
+//    BigInt 경로에서 `RangeError: Division by zero`를 던져 API가 500으로 죽는다
+//    (leaf 쪽 가드는 `lib/tax-engine/tax-utils.ts`에 함께 넣었다).
+//    생산 측(`lib/calc/gift-deemed-api.ts`의 `parseRatio`)은 `{Math.round(pct*100), 10_000}`
+//    으로 언제나 정수를 만들므로, 정수 강제가 정상 입력을 막지 않는다.
+const ratioSchema = z.object({
+  numer: z.number().int().nonnegative(),
+  denom: z.number().int().positive(),
+});
 const mergerShareholderSchema = z.object({
   id: z.string(),
   name: z.string(),
@@ -428,8 +436,9 @@ const specificCorpShareholderSchema = z.object({
   id: z.string(),
   name: z.string(),
   relation: scRelationSchema,
-  shares: z.number().nonnegative(),
-  totalShares: z.number().nonnegative(),
+  // 주식수도 같은 이유로 정수다 — roster 경로의 `shares/totalShares`가 곧 비율 분자·분모다.
+  shares: z.number().int().nonnegative(),
+  totalShares: z.number().int().nonnegative(),
   isDonor: z.boolean(),
   isRelated: z.boolean(),
 });
